Fix dangling reference
